+++ aplicativo/app/src/main/java/com/example/vinic/projetoirrigacao/ValvulaAdapter.java
@@ -33,7 +33,7 @@ public class ValvulaAdapter extends ArrayAdapter {
         TextView val_pulsos= (TextView) convertView.findViewById(R.id.id_pulsos);
 
         val_nome.setText(Integer.toString(valvula.getVal()));
-        val_tempo.setText(Double.toString(valvula.getTempo()));
+        val_tempo.setText(Integer.toString(valvula.getTempo()));
         val_pulsos.setText(Integer.toString(valvula.getPulsos()));
 
         return convertView;
